src/advanced-security-examples/requirements.txt
# Requirements file with known vulnerable dependencies for SonarQube Advanced Security testing
# These dependencies contain known CVEs and should trigger security alerts

# Flask with known vulnerabilities (CVE-2023-30861)
Flask==2.2.2

# Requests with known vulnerabilities (CVE-2023-32681)
requests==2.28.0

# Pillow with known vulnerabilities (CVE-2023-44271, CVE-2023-50447)
Pillow==9.0.0

# PyYAML with known vulnerabilities (CVE-2020-14343)
PyYAML==5.3.1

# Jinja2 with known vulnerabilities (CVE-2020-28493)
Jinja2==2.11.3

# Urllib3 with known vulnerabilities (CVE-2023-43804)
urllib3==1.26.5

# Cryptography with known vulnerabilities (CVE-2023-23931)
cryptography==3.4.8

# Django with known vulnerabilities (CVE-2023-31047)
Django==3.2.13

# SQLAlchemy with known vulnerabilities (CVE-2019-7164)
SQLAlchemy==1.3.0

# Tornado with known vulnerabilities (CVE-2023-28370)
tornado==6.1

# lxml with known vulnerabilities (CVE-2022-2309)
lxml==4.6.5

# httplib2 with known vulnerabilities (CVE-2021-21240)
httplib2==0.18.1

# Note: These versions are intentionally outdated and vulnerable
# Use for testing SonarQube Advanced Security dependency scanning only
# DO NOT use these versions in production environments